Code Optimized,Loading progressbar added
